+++ packages/omniston-sdk/src/logger/Logger.ts
@@ -1,0 +1,6 @@
+export interface Logger {
+  log(message: string, ...extra: unknown[]): void;
+  debug(message: string, ...extra: unknown[]): void;
+  warn(message: string, ...extra: unknown[]): void;
+  error(message: string, ...extra: unknown[]): void;
+}